HLRC: UpdateByQuery API with wait_for_completion being false (#58552)
(cherry picked from commit 291f5bd1b2e889e9447d660e5407f3120cffb1a5)
